Move any_slice_from_array to kani_core
