+++ tests/expected/nondet-slice-u8-oob/main.rs
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: Apache-2.0 OR MIT
 
 // This test checks that Kani reports out-of-bound accesses on a non-det slice
-// created using `kani::slice::any_slice_of_array`
+// created using `kani::any_slice_of_array`
 
 #[kani::proof]
 fn check_out_of_bounds() {
     let arr: [u8; 5] = kani::any();
-    let mut bytes = kani::slice::any_slice_of_array(&arr);
+    let mut bytes = kani::any_slice_of_array(&arr);
     let val = unsafe { *bytes.as_ptr().add(4) };
     kani::assume(val != 0);
     assert_ne!(val, 0);
